requirements.txt
requests==2.31.0
PyGithub==2.1.1
pytest==7.4.3
python-dotenv==1.0.0
typing-extensions==4.8.0
